Update README.md
Absenteeism in the workplace contributes to organizational failure in many aspects and results in reduced productivity among employees. Thus, predicting absenteeism in the workplace using data mining approach helps to understand essential reasons for workplace truancy. In this study, decision tree algorithm was modeled and trained in R Studio and SAS EM to classify reasons for absenteeism into categories defined by ICD. The result of the study identified characteristic factors such as Absenteeism time in hours, Social smoker, and Social drinker as important reasons for absenteeism in the workplace. Also, model accuracy was tested and found that it performed relatively well when compared with results in literature.